requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.5
idna==2.10
joblib==1.0.1
numpy==1.18.4
pandas==0.24.1
pycparser==2.20
python-dateutil==2.8.1
python-firebase==1.2
python-telegram-bot==13.2
pytz==2021.1
requests==2.25.1
scikit-learn==0.20.2
scipy==1.2.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.3